double to float

+++ jacobi.c
@@ -37,9 +37,9 @@ double get_timestamp();
 // Parse command line arguments to set solver parameters
 void parse_arguments(int argc, char *argv[]);
 
-void iterate_jacobi(double *A, double *b, double *x, double *xtmp) {
+void iterate_jacobi(float *A, float *b, float *x, float *xtmp) {
     int row, col;
-    double dot;
+    float dot;
     for (row = 0; row < N; row++)
     {
       dot = 0.0;
@@ -54,14 +54,14 @@ void iterate_jacobi(double *A, double *b, double *x, double *xtmp) {
 
 // Run the Jacobi solver
 // Returns the number of iterations performed
-int run(double *A, double *b, double *x, double *xtmp)
+int run(float *A, float *b, float *x, float *xtmp)
 {
   int itr;
   int row, col;
-  double dot;
-  double diff;
-  double sqdiff;
-  double *ptrtmp;
+  float dot;
+  float diff;
+  float sqdiff;
+  float *ptrtmp;
 
   // Loop until converged or maximum iterations reached
   itr = 0;
@@ -93,10 +93,10 @@ int main(int argc, char *argv[])
 {
   parse_arguments(argc, argv);
 
-  double *A    = malloc(N*N*sizeof(double));
-  double *b    = malloc(N*sizeof(double));
-  double *x    = malloc(N*sizeof(double));
-  double *xtmp = malloc(N*sizeof(double));
+  float *A    = malloc(N*N*sizeof(float));
+  float *b    = malloc(N*sizeof(float));
+  float *x    = malloc(N*sizeof(float));
+  float *xtmp = malloc(N*sizeof(float));
 
   printf(SEPARATOR);
   printf("Matrix size:            %dx%d\n", N, N);
@@ -110,15 +110,15 @@ int main(int argc, char *argv[])
   srand(SEED);
   for (int row = 0; row < N; row++)
   {
-    double rowsum = 0.0;
+    float rowsum = 0.0;
     for (int col = 0; col < N; col++)
     {
-      double value = rand()/(double)RAND_MAX;
+      float value = rand()/(float)RAND_MAX;
       A[row + col*N] = value;
       rowsum += value;
     }
     A[row + row*N] += rowsum;
-    b[row] = rand()/(double)RAND_MAX;
+    b[row] = rand()/(float)RAND_MAX;
     x[row] = 0.0;
   }
 
@@ -128,10 +128,10 @@ int main(int argc, char *argv[])
   double solve_end = get_timestamp();
 
   // Check error of final solution
-  double err = 0.0;
+  float err = 0.0;
   for (int row = 0; row < N; row++)
   {
-    double tmp = 0.0;
+    float tmp = 0.0;
     for (int col = 0; col < N; col++)
     {
       tmp += A[row + col*N] * x[col];
